+++ odoo_dynamic_dashboard/doc/RELEASE_NOTES.md
@@ -1,7 +1,11 @@
 ## Module <odoo_dynamic_dashboard>
 
-#### 29.06.2026
+#### 07.03.2026
 #### Version 19.0.1.0.0
-#### ADD
+##### ADD
+- Initial commit for Odoo Dynamic Dashboard
 
-- Initial commit for AI Dynamic Dashboard Pro
+#### 07.07.2026
+#### Version 19.0.1.0.1
+##### BUG_FIX
+- Fixed the dashboard query issue caused by Odoo's updated SQL handling.